4.21

+++ part4/blog/controllers/blogs.js
@@ -36,6 +36,21 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  console.log(decodedToken)
+
+  const blog = await Blog.findById(request.params.id)
+  console.log(blog)
+
+  if (blog.user.toString() !== decodedToken.id) {
+    return response.status(403).json({ error: 'only the creator of the blog can delete it' })
+  }
+  
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
